Better hotkey support
1. Allows overriding F1--F24, punctuation and special keys (like "home").
2. F1--F24 and special keys work when chat is opened.
3. Saves/loads on the fly via a custom binary file named after the character ID.
4. Setting the item to the same hotkey as it already has clears the hotkey.
5. Original logic is triggered if the hotkey is not set (for example F3).

Potentially we can support letters, but we need to improve the detection
of assigning a spell to a hotkey to preserve "ctrl+u" and similar
vanilla shortcuts.

Doesn't support numbers, because it conflicts with unit groups.

Doesn't support page up and page down, because these buttons are
designated to scrolling the inventory (in another patch), and execution
flow doesn't reach the regular handler.